requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Web interface
flask>=2.3.0
werkzeug>=2.3.0

# C parsing (optional)
pycparser>=2.21
clang>=14.0.0

# Parsing and analysis
lark>=1.1.5

# Test frameworks
unity>=2.5.2
# cmocka>=1.1.5

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Utilities
click>=8.0.0
rich>=12.0.0
typer>=0.7.0 